+++ code/core/src/main/java/org/betonquest/betonquest/id/action/ActionIdentifierFactory.java
@@ -27,6 +27,6 @@ public class ActionIdentifierFactory extends DefaultIdentifierFactory<ActionIden
     public ActionIdentifier parseIdentifier(@Nullable final QuestPackage source, final String input) throws QuestException {
         final Map.Entry<QuestPackage, String> entry = parse(source, input);
         final DefaultActionIdentifier identifier = new DefaultActionIdentifier(entry.getKey(), entry.getValue());
-        return requireInstruction(identifier, DefaultActionIdentifier.ACTION_SECTION);
+        return requireValue(identifier, DefaultActionIdentifier.ACTION_SECTION);
     }
 }